:recycle: 조합 기능 구현 전 페이지 컴포넌트 추상화 및 유틸 파일 폴더 이전
