+++ src/components/CustomCombination/Ingredient.tsx
@@ -1,4 +1,4 @@
-import selectIngredients from '@components/CustomCombination/utils/selectIngredients';
+import selectIngredients from '@utils/customCombination/selectIngredients';
 import styled from '@emotion/styled';
 import { changeRem } from '@styles/mixin';
 import { 인터페이스_꿀조합, 인터페이스_선택된_재료 } from '../../types/ISandwich';
