+++ src/utils/customCombination/selectIngredients.ts
@@ -1,4 +1,4 @@
-import { 인터페이스_꿀조합, 인터페이스_선택된_재료 } from '../../../types/ISandwich';
+import { 인터페이스_꿀조합, 인터페이스_선택된_재료 } from '../../types/ISandwich';
 
 type TProps = {
   나만의_조합: 인터페이스_꿀조합;
